Fixed handling of failed logins, in-reply-to for Twitter, refactored StatusWidget class a bit and prepared it for custom laconi.cas.

+++ qtwitter-app/src/statusedit.cpp
@@ -88,7 +88,7 @@ bool StatusEdit::isStatusClean() const
   return statusClean;
 }
 
-int StatusEdit::getInReplyTo() const
+quint64 StatusEdit::getInReplyTo() const
 {
   return inReplyToId;
 }
